+++ src/Options/Commander/Sequence.hs
@@ -1,0 +1,8 @@
+module Options.Commander.Sequence where
+
+
+-- | The type level program sequencing combinator, taking two program types
+-- and sequencing them one after another.
+data (&) :: k -> * -> *
+infixr 4 &
+
